fix for saveboardconfiguration

+++ src/myFP2ESP/ManagementServer.cpp
@@ -1880,7 +1880,7 @@ void MANAGEMENT_handleadminpg1(void)
       }
       MSrvr_DebugPrint(SETPGTIMESTR);
       MSrvr_DebugPrintln(msg);
-      mySetupData->set_oledpagetime(pgtime);
+      mySetupData->set_oledpagetime((byte)pgtime);
     }
     else
     {
@@ -3037,7 +3037,7 @@ void start_management(void)
   mserver.on("/set",                 MANAGEMENT_handleset);               // generic set function
   mserver.on("/get",                 MANAGEMENT_handleget);               // generic get function
 
-  mserver.on("/config",    HTTP_GET,  MANAGEMENT_config);
+  mserver.on("/config",   HTTP_GET,  MANAGEMENT_config);
   mserver.on("/config",   HTTP_POST, MANAGEMENT_confighandler);
   mserver.on("/showconfig",          MANAGEMENT_showboardconfig);
   mserver.on("/custombrd",           MANAGEMENT_custombrd);
